feat(helpers): add generic error handlers for read write and http
Add generic helper functions (closures) that can take a value with a generic type
and an error. If the error is not nil the closure funcition handles the error
and panic in the case of read write or tryes to do a http error and logs the error
in the case of an server error.